validation externe et interne

+++ header/init.h
@@ -1,7 +1,7 @@
 #ifndef _INIT_H_
 #define _INIT_H_
 LABYRINTHE init_labyrinthe(FILE* FIC,char* filename);
-void LIB_MAT(int NB_L,int** MAT);
+void LIB_MAT(int NB_L,unsigned short int** MAT);
 void remplissage_Map(LABYRINTHE labyrinthe,FILE* FIC,char* filename);
 
 #endif
